Move IServiceCollection.AddAllActuators into subdirectory, document
